+++ src/web/BadRequestHttpException.php
@@ -8,7 +8,7 @@
 
 namespace yii\web;
 
-use Exception;
+use Throwable;
 
 /**
  * BadRequestHttpException represents a "Bad Request" HTTP exception with status code 400.
@@ -31,9 +31,9 @@ class BadRequestHttpException extends HttpException
      *
      * @param string $message error message
      * @param int $code error code
-     * @param Exception $previous The previous exception used for the exception chaining.
+     * @param Throwable|null $previous The previous exception used for the exception chaining.
      */
-    public function __construct($message = null, $code = 0, Exception $previous = null)
+    public function __construct($message = null, $code = 0, $previous = null)
     {
         parent::__construct(400, $message, $code, $previous);
     }